laid the groundwork for a cli version
