+++ coding_assistant/clipy.py
@@ -1,0 +1,9 @@
+# I don't want to be sued my MS so I'll use a single 'p'
+clipy = """
+ _-_  | /
+/_  \\ |/
+(o)(o)
+| | |
+| \\/ /
+\\    |
+ ¯--¯"""
